requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
distlib==0.3.8
ecdsa==0.19.0
fastapi==0.115.8
filelock==3.13.4
greenlet==3.1.1
h11==0.14.0
idna==3.10
passlib==1.7.4
platformdirs==4.2.0
py-dotenv==0.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.4.0
rsa==4.9
setuptools==69.2.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
virtualenv==20.25.1
wheel==0.43.0
